graphics upgrade
I made the start game button but only the graphics. for the x-button i
made also mouse-over and mouse-down versions though they won’t be
needed so much in mobile version. Also a splash screen that might just
work as an animated object staying still for 2 secs when we start the
game, i’ll look into that.