new animated sprite sheets
i got the other conversion and finally the split animations. But the
lila split is oriented the same way the as the previously used
temporary/test split. In other new lila’s sprite sheets it is looking
the other way, so hope this works or can be mirrored at some state, as
it saved me lots of time this way